requirements.txt
fastapi>=0.110.3
uvicorn>=0.29.0
pyjwt>=2.8.0
python-dotenv>=1.0.1
pydantic>=2.7.1
requests>=2.31.0
email-validator>=2.1.1
psycopg2-binary~=2.9.6
bcrypt~=4.0.1
sqlalchemy>=2.0.29
cryptography>=42.0.5
python-multipart~=0.0.6
